requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.20
Django==1.8.15
django-appconf==1.0.1
django-compressor==2.0
django-recaptcha==1.0.5
django-wysiwyg-redactor==0.4.9
kombu==3.0.33
python-dateutil==2.5.0
pytz==2015.7
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
uWSGI==2.0.14
